resume-analyzer/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pymongo
pdfminer.six
python-docx
